Подключает моки в файл index.tsx и передает данные через пропсы в компонент `App`
