+++ src/components/layout/layout.tsx
@@ -6,7 +6,11 @@ import Footer from '../../pages/favorites/components/footer.tsx';
 import { getLayoutState } from '../../utils/common.ts';
 import { getAuthorizationStatus } from '../../mock/authorizationStatus.ts';
 
-export default function Layout(): JSX.Element {
+type LayoutProps = {
+  favoriteCount: number;
+}
+
+export default function Layout({favoriteCount}: LayoutProps): JSX.Element {
 
   const {pathname} = useLocation();
   const {rootClassName, linkClassName, title} = getLayoutState(pathname as AppRoute);
@@ -17,7 +21,7 @@ export default function Layout(): JSX.Element {
       <Helmet>
         <title>{`6 cities${title}`}</title>
       </Helmet>
-      <Header linkClassName={linkClassName} authorizationStatus={authorizationStatus} pathname={pathname as AppRoute}/>
+      <Header linkClassName={linkClassName} authorizationStatus={authorizationStatus} pathname={pathname as AppRoute} favoriteCount={favoriteCount} />
       <Outlet />
       {pathname as AppRoute === AppRoute.Favorites && <Footer />}
     </div>
