+++ src/index.tsx
@@ -1,7 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/app/app';
-import {Setting} from './const';
+import {Setting} from './const.ts';
+import {offer, offers, offersNear, favorites, favoriteCount } from './mock/offer.ts';
+import { reviews } from './mock/review.ts';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -9,6 +11,6 @@ const root = ReactDOM.createRoot(
 
 root.render(
   <React.StrictMode>
-    <App offersCount = {Setting.OffersCount} />
+    <App offersCount = {Setting.OffersCount} offer={offer} offers={offers} offersNear={offersNear} favorites={favorites} favoriteCount={favoriteCount} reviews={reviews} />
   </React.StrictMode>,
 );
