+++ src/pages/favorites/favorites.tsx
@@ -1,18 +1,13 @@
-import { favorites } from '../../const.ts';
 import { TypeOfferCard } from '../../types.ts';
 import LocationsItemLink from '../../components/ui/locations-item-link.tsx';
 import PlaceCard from '../../components/place-card.tsx';
-
-const favoritesGroupped: { [city: string]: TypeOfferCard[] } = favorites.reduce((result, favorite) => ({
-  ...result, [favorite.city.name]: [...(result[favorite.city.name] || []), favorite]
-}), {} as { [city: string]: TypeOfferCard[] });
 
 type FavoritesInCityProps = {
   city: string;
   offers: TypeOfferCard[];
 }
 
-function FavoritesInCity({city, offers}: FavoritesInCityProps): JSX.Element {
+function FavoritesInCity({city, offers }: FavoritesInCityProps): JSX.Element {
   return (
     <li className="favorites__locations-items">
       <div className="favorites__locations locations locations--current">
@@ -32,7 +27,16 @@ function FavoritesInCity({city, offers}: FavoritesInCityProps): JSX.Element {
   );
 }
 
-export default function Favorites(): JSX.Element {
+type FavoritesProps = {
+  favorites: TypeOfferCard[];
+}
+
+export default function Favorites({favorites}: FavoritesProps): JSX.Element {
+
+  const favoritesGroupped: { [city: string]: TypeOfferCard[] } = favorites.reduce((result, favorite) => ({
+    ...result, [favorite.city.name]: [...(result[favorite.city.name] || []), favorite]
+  }), {} as { [city: string]: TypeOfferCard[] });
+
 
   const cities = Object.keys(favoritesGroupped);
